# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pilates Studio Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PilatesClassLevel =
  (typeof PilatesClassLevel)[keyof typeof PilatesClassLevel];

export const PilatesClassLevel = {
  Principiante: "Principiante",
  Intermedio: "Intermedio",
  Avanzado: "Avanzado",
} as const;

export type PilatesClassType =
  (typeof PilatesClassType)[keyof typeof PilatesClassType];

export const PilatesClassType = {
  Reformer: "Reformer",
  Mat: "Mat",
  Privada: "Privada",
} as const;

export type PilatesClassStatus =
  (typeof PilatesClassStatus)[keyof typeof PilatesClassStatus];

export const PilatesClassStatus = {
  Activa: "Activa",
  Cancelada: "Cancelada",
  Completa: "Completa",
} as const;

export interface PilatesClass {
  id: number;
  name: string;
  instructor: string;
  instructorId: number;
  time: string;
  /** Duration in minutes */
  duration: number;
  capacity: number;
  enrolled: number;
  level: PilatesClassLevel;
  type: PilatesClassType;
  status: PilatesClassStatus;
  dayOfWeek: string;
  date: string;
}

export type CreateClassBodyLevel =
  (typeof CreateClassBodyLevel)[keyof typeof CreateClassBodyLevel];

export const CreateClassBodyLevel = {
  Principiante: "Principiante",
  Intermedio: "Intermedio",
  Avanzado: "Avanzado",
} as const;

export type CreateClassBodyType =
  (typeof CreateClassBodyType)[keyof typeof CreateClassBodyType];

export const CreateClassBodyType = {
  Reformer: "Reformer",
  Mat: "Mat",
  Privada: "Privada",
} as const;

export interface CreateClassBody {
  name: string;
  instructorId: number;
  time: string;
  duration: number;
  capacity: number;
  level: CreateClassBodyLevel;
  type: CreateClassBodyType;
  dayOfWeek: string;
  date: string;
}

export type ClientPlan = (typeof ClientPlan)[keyof typeof ClientPlan];

export const ClientPlan = {
  Mensual: "Mensual",
  Trimestral: "Trimestral",
  Anual: "Anual",
  Por_Clase: "Por Clase",
} as const;

export interface Client {
  id: number;
  name: string;
  phone: string;
  email: string;
  plan: ClientPlan;
  classesRemaining: number;
  notes?: string;
  createdAt: string;
}

export type CreateClientBodyPlan =
  (typeof CreateClientBodyPlan)[keyof typeof CreateClientBodyPlan];

export const CreateClientBodyPlan = {
  Mensual: "Mensual",
  Trimestral: "Trimestral",
  Anual: "Anual",
  Por_Clase: "Por Clase",
} as const;

export interface CreateClientBody {
  name: string;
  phone: string;
  email: string;
  plan: CreateClientBodyPlan;
  classesRemaining: number;
  notes?: string;
}

export interface AttendanceRecord {
  id: number;
  classId: number;
  className: string;
  date: string;
  attended: boolean;
}

export interface Instructor {
  id: number;
  name: string;
  email: string;
  phone: string;
  specialties: string[];
}

export interface CreateInstructorBody {
  name: string;
  email: string;
  phone: string;
  specialties: string[];
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  Confirmada: "Confirmada",
  Pendiente: "Pendiente",
  Cancelada: "Cancelada",
} as const;

export interface Reservation {
  id: number;
  clientId: number;
  clientName: string;
  classId: number;
  className: string;
  date: string;
  status: ReservationStatus;
}

export interface CreateReservationBody {
  clientId: number;
  classId: number;
  date: string;
}

export interface DashboardSummary {
  todayClassesCount: number;
  activeClientsCount: number;
  pendingReservationsCount: number;
  availableSpotsCount: number;
}