# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pilates Studio Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ClientPlan } from "./clientPlan";

export interface Client {
  id: number;
  name: string;
  phone: string;
  email: string;
  plan: ClientPlan;
  classesRemaining: number;
  notes?: string;
  createdAt: string;
}