# lib/api-zod/src/generated/types/createClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pilates Studio Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateClientBodyPlan } from "./createClientBodyPlan";

export interface CreateClientBody {
  name: string;
  phone: string;
  email: string;
  plan: CreateClientBodyPlan;
  classesRemaining: number;
  notes?: string;
}