# lib/api-zod/src/generated/types/createInstructorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pilates Studio Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInstructorBody {
  name: string;
  email: string;
  phone: string;
  specialties: string[];
}